Cyber Backend (With Routing)/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
motor
pymongo
pytz
redis
PyJWT
pandas
numpy
email-validator